Add vitest coverage to the CLI package
Coverage could not see the e2e suite: it spawned the bundled bin as a
child process, leaving the whole command layer unmeasured. Extract the
program construction into createProgram() so the tests drive it in this
process instead, and mirror cli.ts's error handling in the helper so
exit codes and stderr keep their meaning.

Merge the two vitest configs into unit and e2e projects so a single run
reports both suites together. The e2e files share one anvil node and
one funded account, so they run serially to avoid racing on nonces and
allowances.

Also split swap.test.ts into mint and approve, matching the per-command
layout of the other e2e files, leaving swap.test.ts with just the
cross-command flow.
