+++ packages/cli/package.json
@@ -18,8 +18,9 @@
     "prepack": "cp ../../LICENSE LICENSE",
     "postpack": "rm -f LICENSE",
     "prepublishOnly": "pnpm clean && pnpm bundle",
-    "test": "vitest run",
-    "test:e2e": "pnpm bundle && vitest run --config vitest.e2e.config.ts"
+    "test": "vitest run --project unit",
+    "test:coverage": "vitest run --coverage",
+    "test:e2e": "vitest run --project e2e"
   },
   "dependencies": {
     "@vetro-protocol/gateway": "workspace:*",
@@ -32,6 +33,7 @@
     "@hemilabs/anvil-fork-setup": "2.0.0",
     "@tsconfig/node24": "24.0.4",
     "@types/node": "24.1.0",
+    "@vitest/coverage-v8": "4.1.9",
     "esbuild": "0.27.3",
     "typescript": "5.9.3",
     "vitest": "4.1.9"
